e2e: test filter by user w/ playwright. Follow up
